+++ app/web/features/auth/password/index.ts
@@ -1,3 +1,3 @@
 export { default as ChangePassword } from "./ChangePassword";
 export { default as ResetPassword } from "./ResetPassword";
-export { default as CompletePasswordReset } from "./CompletePasswordReset";
+export { default as CompleteResetPassword } from "./CompleteResetPassword";
