+++ app/web/pages/complete-password-reset.tsx
@@ -1,5 +1,5 @@
 import { appGetLayout } from "components/AppRoute";
-import { CompletePasswordReset } from "features/auth/password";
+import { CompleteResetPassword } from "features/auth/password";
 import { GetStaticProps } from "next";
 import nextI18nextConfig from "next-i18next.config";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
@@ -15,7 +15,7 @@ export const getStaticProps: GetStaticProps = async ({ locale }) => ({
 });
 
 export default function CompletePasswordResetPage() {
-  return <CompletePasswordReset />;
+  return <CompleteResetPassword />;
 }
 
 CompletePasswordResetPage.getLayout = appGetLayout({
